separate nft_req and nft component
